+++ src/main/java/net/countercraft/movecraft/warfare/features/assault/commands/AssaultRepairCommand.java
@@ -2,7 +2,7 @@ package net.countercraft.movecraft.warfare.features.assault.commands;
 
 import net.countercraft.movecraft.warfare.MovecraftWarfare;
 import net.countercraft.movecraft.warfare.config.Config;
-import net.countercraft.movecraft.warfare.features.assault.WarfareRepair;
+import net.countercraft.movecraft.warfare.features.assault.AssaultRepair;
 import net.countercraft.movecraft.warfare.features.assault.events.AssaultBroadcastEvent;
 import net.countercraft.movecraft.warfare.localisation.I18nSupport;
 import net.countercraft.movecraft.worldguard.MovecraftWorldGuard;
