+++ src/main/java/net/countercraft/movecraft/warfare/features/assault/tasks/AssaultTask.java
@@ -4,7 +4,7 @@ import net.countercraft.movecraft.warfare.localisation.I18nSupport;
 import net.countercraft.movecraft.warfare.MovecraftWarfare;
 import net.countercraft.movecraft.warfare.config.Config;
 import net.countercraft.movecraft.warfare.features.assault.Assault;
-import net.countercraft.movecraft.warfare.features.assault.WarfareRepair;
+import net.countercraft.movecraft.warfare.features.assault.AssaultRepair;
 import net.countercraft.movecraft.warfare.features.assault.events.AssaultBroadcastEvent;
 import net.countercraft.movecraft.warfare.features.assault.events.AssaultLoseEvent;
 import net.countercraft.movecraft.warfare.features.assault.events.AssaultWinEvent;
