+++ src/main/java/net/countercraft/movecraft/warfare/features/assault/tasks/ChunkRepairTask.java
@@ -3,7 +3,7 @@ package net.countercraft.movecraft.warfare.features.assault.tasks;
 import net.countercraft.movecraft.MovecraftLocation;
 import net.countercraft.movecraft.warfare.MovecraftWarfare;
 import net.countercraft.movecraft.warfare.config.Config;
-import net.countercraft.movecraft.warfare.features.assault.WarfareRepair;
+import net.countercraft.movecraft.warfare.features.assault.AssaultRepair;
 import net.countercraft.movecraft.warfare.features.assault.events.AssaultBroadcastEvent;
 import net.countercraft.movecraft.warfare.localisation.I18nSupport;
